# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: DINO API
  version: 1.0.0
  description: |
    # 🦕 DINO - Digital Nomad Travel Management API
    
    REST API for managing digital nomad travel records, visa compliance, and Schengen area calculations.
    
    ## Features
    - ✅ Trip management with CRUD operations
    - ✅ Real-time Schengen 90/180 rule calculations
    - ✅ Google Calendar integration
    - ✅ Gmail email analysis for travel extraction
    - ✅ Statistical analytics and reporting
    - ✅ Data export/import capabilities
    
    ## Authentication
    Uses NextAuth.js with Google OAuth 2.0 and session-based authentication.
    
    ## Rate Limiting
    - General API: 100 requests per 15 minutes
    - Mutations: 20 requests per 15 minutes  
    - Authentication: 5 requests per 15 minutes
    - Data Export: 10 requests per hour
    
  contact:
    name: DINO API Support
    url: https://dinoapp.net/support
    email: support@dinoapp.net
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://dinoapp.net/terms
servers:
  - url: https://dinoapp.net/api
    description: Production server
  - url: https://staging.dinoapp.net/api
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
security:
  - cookieAuth: []
  - csrfToken: []
paths:
...
  /trips:
    get:
      tags: [Trips]
      summary: Get user trips
      description: |
        Retrieve paginated list of user's travel records with optional filtering.
        
        ### Query Parameters
        - `country`: Filter by country name or ISO code
        - `year`: Filter by year (e.g., 2024)
        - `visa_type`: Filter by visa type
        - `page`: Page number (default: 1)
        - `limit`: Items per page (default: 20, max: 100)
        - `sort`: Sort field and direction (e.g. "entry_date:desc")
      parameters:
        - name: country
          in: query
          description: Filter trips by country name or ISO code
          schema:
            type: string
            example: "France"
        - name: year
          in: query
          description: Filter trips by year
          schema:
            type: integer
            example: 2024
        - name: visa_type
          in: query
          description: Filter by visa type
          schema:
            type: string
            enum: [Tourist, Business, Student, Working Holiday, Transit, Medical, Family Visit]
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: sort
          in: query
          description: Sort field and direction
          schema:
            type: string
            example: "entry_date:desc"
      responses:
        '200':
          description: List of user trips
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Trips]
      summary: Create new trip
      description: |
        Create a new travel record with automatic visa validation and Schengen calculations.
        
        ### Validation Rules
        - Entry date cannot be in the future beyond 1 year
        - Exit date must be after entry date (if provided)
        - Visa type must match passport country eligibility
        - Country must be a valid ISO country code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreateRequest'
            examples:
              tourist_trip:
                summary: Tourist trip to France
                value:
                  country: "France"
                  entryDate: "2024-06-01"
                  exitDate: "2024-06-15"
                  visaType: "Tourist"
                  maxDays: 90
                  passportCountry: "US"
                  notes: "Summer vacation in Paris"
              business_trip:
                summary: Business trip to Germany
                value:
                  country: "Germany"
                  entryDate: "2024-07-10"
                  exitDate: "2024-07-12"
                  visaType: "Business"
                  maxDays: 90
                  passportCountry: "US"
                  notes: "Tech conference in Berlin"
      responses:
        '201':
          description: Trip created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Trip conflicts with existing record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /schengen:
    get:
      tags: [Schengen Calculator]
      summary: Get current Schengen status
      description: |
        Calculate the user's current Schengen Area compliance status based on the 90/180 day rule.
        
        ### Calculation Method
        - Uses rolling 180-day window
        - Counts all days spent in Schengen countries
        - Provides warnings for approaching limits
        - Calculates next available entry dates
      responses:
        '200':
          description: Current Schengen status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchengenStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
      description: NextAuth.js session cookie
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF protection token
    adminRole:
      type: http
      scheme: bearer
      description: Admin role verification
  schemas:
...
    Trip:
      type: object
      properties:
        id:
          type: string
          description: Unique trip identifier
          example: "clp123456789abcdef"
        userId:
          type: string
          description: User identifier
          example: "usr_987654321xyz"
        country:
          type: string
          description: Destination country name
          example: "France"
        countryCode:
          type: string
          description: ISO country code
          example: "FR"
        entryDate:
          type: string
          format: date-time
          description: Entry date and time
          example: "2024-06-01T00:00:00.000Z"
        exitDate:
          type: string
          format: date-time
          nullable: true
          description: Exit date and time (null if still in country)
          example: "2024-06-15T23:59:59.999Z"
        visaType:
          type: string
          enum: [Tourist, Business, Student, Working Holiday, Transit, Medical, Family Visit]
          description: Type of visa or entry permission
          example: "Tourist"
        maxDays:
          type: integer
          minimum: 1
          maximum: 365
          description: Maximum allowed days for this visa type
          example: 90
        passportCountry:
          type: string
          description: Passport issuing country
          example: "US"
        notes:
          type: string
          nullable: true
          maxLength: 500
          description: User notes about the trip
          example: "Summer vacation in Paris and Nice"
        createdAt:
          type: string
          format: date-time
          description: Record creation timestamp
          example: "2024-05-15T10:30:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: "2024-05-16T14:20:00.000Z"
      required: [id, userId, country, entryDate, visaType, maxDays, passportCountry]
...
    TripCreateRequest:
      type: object
      properties:
        country:
          type: string
          minLength: 2
          maxLength: 100
          description: Destination country name
          example: "France"
        entryDate:
          type: string
          format: date
          description: Entry date (YYYY-MM-DD)
          example: "2024-06-01"
        exitDate:
          type: string
          format: date
          nullable: true
          description: Exit date (YYYY-MM-DD), optional for ongoing trips
          example: "2024-06-15"
        visaType:
          type: string
          enum: [Tourist, Business, Student, Working Holiday, Transit, Medical, Family Visit]
          description: Type of visa or entry permission
          example: "Tourist"
        maxDays:
          type: integer
          minimum: 1
          maximum: 365
          description: Maximum allowed days for this visa type
          example: 90
        passportCountry:
          type: string
          description: Passport issuing country
          example: "US"
        notes:
          type: string
          maxLength: 500
          nullable: true
          description: Optional notes about the trip
          example: "Summer vacation in Paris"
      required: [country, entryDate, visaType, maxDays, passportCountry]
...
    TripResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            trip:
              $ref: '#/components/schemas/Trip'
        message:
          type: string
          example: "Trip retrieved successfully"
    TripListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            trips:
              type: array
              items:
                $ref: '#/components/schemas/Trip'
            pagination:
              type: object
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 20
                total:
                  type: integer
                  example: 45
                totalPages:
                  type: integer
                  example: 3
    SchengenStatusResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            status:
              type: object
              properties:
                daysUsed:
                  type: integer
                  description: Days used in current 180-day period
                  example: 45
                daysRemaining:
                  type: integer
                  description: Days remaining in current period
                  example: 45
                earliestExitDate:
                  type: string
                  format: date-time
                  nullable: true
                  description: Earliest required exit date to stay compliant
                nextAvailableEntry:
                  type: string
                  format: date-time
                  description: Next date when 90 days will be available
                currentlyInSchengen:
                  type: boolean
                  description: Whether user is currently in Schengen area
                periodStart:
                  type: string
                  format: date-time
                  description: Start of current 180-day calculation period
                periodEnd:
                  type: string
                  format: date-time
                  description: End of current 180-day calculation period
            warnings:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    enum: [APPROACHING_LIMIT, EXCEEDED_LIMIT, INVALID_STAY]
                  message:
                    type: string
                  severity:
                    type: string
                    enum: [info, warning, error]
                  daysRemaining:
                    type: integer
                    nullable: true
            trips:
              type: array
              items:
                type: object
                properties:
                  country:
                    type: string
                  entryDate:
                    type: string
                    format: date-time
                  exitDate:
                    type: string
                    format: date-time
                    nullable: true
                  daysInSchengen:
                    type: integer
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              enum: [BAD_REQUEST, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, CONFLICT, VALIDATION_ERROR, RATE_LIMIT_ERROR, INTERNAL_SERVER_ERROR, DATABASE_ERROR, EXTERNAL_SERVICE_ERROR]
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              nullable: true
              description: Additional error details (validation errors, etc.)
            requestId:
              type: string
              description: Unique request identifier for debugging
            timestamp:
              type: string
              format: date-time
              description: Error timestamp
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            validation_error:
              value:
                success: false
                error:
                  code: VALIDATION_ERROR
                  message: "Validation failed"
                  details:
                    country: ["Country is required"]
                    entryDate: ["Invalid date format"]
                  requestId: "req_123456789_abc"
                  timestamp: "2024-01-01T00:00:00.000Z"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            not_authenticated:
              value:
                success: false
                error:
                  code: UNAUTHORIZED
                  message: "Authentication required"
                  requestId: "req_123456789_def"
                  timestamp: "2024-01-01T00:00:00.000Z"
...